Catch version exists error
