+++ ml_enabler/models/utils.py
@@ -1,5 +1,8 @@
 import datetime
 
+class VersionExists(Exception):
+    pass
+
 class NotFound(Exception):
     """ Custom exception to indicate model not found in database"""
     pass
